+++ CHANGELOG.md
@@ -23,6 +23,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 - Drop support for `mentions` parameter in `/api/v1/statuses` form data.
 
+### Fixed
+
+- Fix clearing of attachments during post update.
+
 ## [2.1.0] - 2023-10-10
 
 ### Added